--- newlogin.css
@@ -1,16 +1,0 @@
-body{
-    font-size: 40px;
-    align-self: center;
-}
-.logo-img{                  /*클래스를 불러올 때는 클래스 이름 앞에 .을 붙여서 가져온다.*/
-    font-size: 90px;
-}
-form{
-    border-radius: 30px solid blue;
-    padding: 10px;
-    margin: 10px;
-    align-self: center;
-}
-div{
-    font-size: 40px;
-}
