--- newsignup.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-    <title>회원가입 페이지</title>
-    <meta charset="UTF-8">
-    <link rel="stylesheet" href="newsignup.css">
-</head>
-<body>
-    <img src="" alt="WealthWise">       <!--img 태그는 이미지를 불러오는 태그, src는 삽입하려는 이미지의 경로(나중에 로고이미지 경로 넣으면 된다.)를 삽입하는 태그, alt는 이미지 경로가 연결되지 않을 경우의 대체 택스트이다.-->
-    <form>
-        <div>
-            <input type="text" placeholder="이름">      <!--placeholder는 사용자의 입력란에 무엇을 입력해야하는지 알려주는 말을 넣는 태그이다. 비슷한 것으로 value가 있는데, 이는 사용자가 수정하지 않으면 코드에 지정된 value값으로 내용이 입력되는 것이다.-->
-        </div>
-        <div>
-            <input type="text" placeholder="ID">
-        </div>
-        <div>
-            <input type="text" placeholder="비밀번호">
-        </div>
-        <div>
-            <input type="text" placeholder="비밀번호 확인">
-        </div>
-        <div>
-            <button type="submit">회원가입</button>
-        </div>
-    </form>
-</body>
-</html>
