--- zeroback/src/component/Sign in Page/Sign in and Sign up.css
@@ -1,63 +1,0 @@
-/* 전체 양식을 감싸는 div에 스타일을 적용합니다. */
-.Form-wrapper {
-    max-width: 400px; /* 양식의 최대 너비를 설정합니다. */
-    margin: 0 auto; /* 가운데 정렬을 위해 왼쪽과 오른쪽 여백을 자동 조정합니다. */
-    padding: 20px; /* 내부 여백을 추가하여 내용이 양식의 가장자리에 붙지 않도록 합니다. */
-    border: 1px solid #ccc; /* 양식 주위에 회색 테두리를 추가합니다. */
-    border-radius: 5px; /* 양식의 모서리를 둥글게 만듭니다. */
-}
-
-/* 로고 컨테이너에 스타일을 적용합니다. */
-#logo-container {
-    text-align: center; /* 로고를 가운데 정렬합니다. */
-    margin-bottom: 20px; /* 로고 아래 여백을 추가합니다. */
-}
-
-/* 버튼에 공통적으로 적용되는 스타일을 지정합니다. */
-.btn {
-    width: 100%; /* 버튼을 양식 너비에 맞게 설정합니다. */
-    padding: 10px; /* 내부 여백을 추가하여 버튼 내용이 너무 가깝게 붙지 않도록 합니다. */
-    margin-bottom: 10px; /* 버튼 아래 여백을 추가합니다. */
-    border: none; /* 버튼 주위에 테두리를 없앱니다. */
-    border-radius: 5px; /* 버튼의 모서리를 둥글게 만듭니다. */
-    background-color: #007bff; /* 버튼 배경색을 파란색으로 설정합니다. */
-    color: #fff; /* 버튼 텍스트 색상을 흰색으로 설정합니다. */
-    cursor: pointer; /* 커서를 포인터로 변경하여 버튼이 클릭 가능한 것임을 나타냅니다. */
-}
-
-/* 버튼에 호버 효과를 추가합니다. */
-.btn:hover {
-    background-color: #0056b3; /* 호버 시 버튼 배경색을 어두운 파란색으로 변경합니다. */
-}
-
-/* 입력란에 스타일을 적용합니다. */
-input[type="text"],
-input[type="password"] {
-    width: 100%; /* 입력란을 양식 너비에 맞게 설정합니다. */
-    padding: 10px; /* 내부 여백을 추가하여 내용이 입력란의 가장자리에 붙지 않도록 합니다. */
-    margin-bottom: 10px; /* 입력란 아래 여백을 추가합니다. */
-    border: 1px solid #ccc; /* 입력란 주위에 회색 테두리를 추가합니다. */
-    border-radius: 5px; /* 입력란의 모서리를 둥글게 만듭니다. */
-}
-
-/* 체크박스 레이블에 스타일을 적용합니다. */
-label {
-    display: block; /* 레이블을 블록 수준 요소로 만들어서 레이블 위아래 여백을 추가합니다. */
-    margin-bottom: 10px; /* 레이블 아래 여백을 추가합니다. */
-}
-
-/* 제출 버튼에 공통적으로 적용되는 스타일을 지정합니다. */
-.submit-btn {
-    width: 100%; /* 제출 버튼을 양식 너비에 맞게 설정합니다. */
-    padding: 10px; /* 내부 여백을 추가하여 내용이 버튼의 가장자리에 붙지 않도록 합니다. */
-    border: none; /* 버튼 주위에 테두리를 없앱니다. */
-    border-radius: 5px; /* 버튼의 모서리를 둥글게 만듭니다. */
-    background-color: #28a745; /* 제출 버튼의 배경색을 초록색으로 설정합니다. */
-    color: #fff; /* 버튼 텍스트 색상을 흰색으로 설정합니다. */
-    cursor: pointer; /* 커서를 포인터로 변경하여 버튼이 클릭 가능한 것임을 나타냅니다. */
-}
-
-/* 제출 버튼에 호버 효과를 추가합니다. */
-.submit-btn:hover {
-    background-color: #218838; /* 호버 시 버튼 배경색을 어두운 초록색으로 변경합니다. */
-}
