--- zeroback/src/component/Sign in Page/Sign in and Sign up.html
@@ -1,64 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Authentication Form</title>
-    <link rel="stylesheet" href="Sign in and Sign up.css">
-    <style>
-        body {
-            margin: 0;
-            padding: 0;
-            background-image: url('./pulse-trace-glowing-computer-monitor-generated-by-ai.jpg');
-            background-size: cover;
-            background-position: center;
-            font-family: Arial, sans-serif; /* Set font family */
-        }
-    </style>
-</head>
-<body>
-    <div class="Form-wrapper">
-        <div id="logo-container">
-            <!-- Placeholder for logo image -->
-        </div>
-
-        <button class="btn" onclick="showLoginForm()">Sign In</button>
-        <button class="btn" onclick="showRegistrationForm()">Sign Up</button>
-
-        <div id="login-form" style="display: none;">
-            <h2>Sign In</h2>
-            <form method="post" action="server-url">
-                <input type="text" name="username" placeholder="Username">
-                <input type="password" name="userPassword" placeholder="Password">
-                <label for="remember-check">
-                    <input type="checkbox" id="remember-check">Remember me
-                </label>
-                <input type="submit" class="submit-btn" value="Sign In">
-            </form>
-        </div>
-
-        <div id="registration-form" style="display: none;">
-            <h2>Sign Up</h2>
-            <form method="post" action="server-url">
-                <input type="text" name="name" placeholder="Name">
-                <input type="text" name="username" placeholder="Username">
-                <input type="password" name="userPassword" placeholder="Password">
-                <input type="password" name="confirmPassword" placeholder="Confirm Password">
-                <input type="submit" class="submit-btn" value="Sign Up">
-            </form>
-        </div>
-    </div>
-
-    <script>
-        function showLoginForm() {
-            document.getElementById('login-form').style.display = 'block';
-            document.getElementById('registration-form').style.display = 'none';
-        }
-
-        function showRegistrationForm() {
-            document.getElementById('registration-form').style.display = 'block';
-            document.getElementById('login-form').style.display = 'none';
-        }
-    </script>
-</body>
-</html>
