--- zeroback/src/component/Sign in Page/Sign in.html
@@ -1,13 +1,0 @@
-<body>
-    <div class="Sign In-wrapper">
-        <h2>Sign In</h2>
-        <form method="post" action="서버의url" id="Sign In-form">
-            <input type="text" name="username" placeholder="ID">
-            <input type="password" name="userPassword" placeholder="Password">
-            <label for="remember-check">
-                <input type="checkbox" id="remember-check">아이디 저장하기
-            </label>
-            <input type="submit" value="Sign In">
-        </form>
-    </div>
-</body>
