--- zeroback/src/component/Sign in Page/Sign_in_Sign_up.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-    <meta charset="UTF-8">
-    <title>Document</title>
-    <link rel="stylesheet" href="Sign_in_Sign_up.css">
-</head>
-<body>
-    <section class="SignIn-form">
-        <h1>WealthWise</h1>
-        <form action="">
-            <div class="int-area">
-                <input type="text" name="id" id="id"
-                autocomplete="off" required>
-                <label for="id">USER NAME</label>
-            </div>
-            <div class="int-area">
-                <input type="password" name="pw" id="pw"
-                autocomplete="off" required>
-                <label for="pw">PASSWORD</label>
-            </div>
-            <div class="btn-area">
-                <button id="btn"
-                type="submit">SIGNIN</button>
-            </div>
-        </form>
-        <div class="caption">
-            <a href="">Forgot Password?</a>
-        </div>
-    </section>
-</body>
-</html>
