--- zeroback/src/component/board.css
@@ -1,140 +1,0 @@
-@import url('https://fonts.googleapis.com/css?family=Inter');
-
-#NavBar {
-    margin-bottom: 30px;
-}
-
-.links {
-    display: flex;
-    align-items: end;
-    flex-direction: column;
-}
-
-.link {
-    margin-right: 3%;
-    font-style: normal;
-    font-weight: 100;
-    font-size: 30px;
-    line-height: 73px;
-    text-decoration: none;
-
-    color: #000000;
-
-    text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-}
-
-.link:hover {
-    cursor: pointer;
-    transform: scale(110%);
-    transition: transform 0.5s cubic-bezier(0.1,1.0,1.0,1.0), opacity 0.55s ease;
-}
-
-#Board {
-    width: 94%;
-    margin-left: 3%;
-    margin-right: 3%;
-    margin-bottom: 2%;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-}
-
-#MainBoard {
-    width: 63%;
-    height: 1000px;
-
-    background: #F9F9F9;
-    border-radius: 30px;
-}
-
-#MainBoardMenu {
-    background-color: #F9F9F9;  
-    display: flex;
-    justify-content: space-between;
-    border-radius: 30px;
-}
-
-#MainBoardMenu li {
-    list-style: none;
-    background-color: #F9F9F9;
-    font-family:fantasy;
-    font-style: normal;
-    font-weight: 100;
-    width: 90px;
-    color: #000000;
-}
-
-#MainBoardMenu p {
-    text-decoration: none;
-}
-
-#MainBoardPostLines  {
-    justify-content: space-between;
-    display: flex;
-    flex-direction: column;
-    position: absolute;
-    width: 95%;
-
-    margin-left: auto;
-    margin-right: auto;
-    left: 0;
-    right: 0;
-}
-
-.MainPostLine {
-    margin-top: 4%;
-    border: 1.5px solid #E9E9E9;
-}
-
-#SubBoard {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-}
-
-.SubBoards {
-    width: 480pX;
-    height: 445px;
-    background: #F9F9F9;
-    border-radius: 30px;
-}
-
-.SubBoards p {
-    background-color: #F9F9F9;
-    border-radius: 30px;
-    text-align: center;
-}
-
-#buttons {
-    display: flex;
-    justify-content: space-between;
-}
-
-#buttons button {
-    cursor: pointer;
-    background: #F9F9F9;
-    height: 45px;
-    border-radius: 40px;
-    border: none;
-}
-
-#search_img {
-    background-color: #F9F9F9;
-}
-
-#write {
-    width: 160px;
-}
-
-#scrap_post_manage {
-    width: 210px;
-}
-
-#search_button {
-    width: 45px;
-}
-
-#search {
-    background-color: #F9F9F9;
-    width: 28px;
-}
