--- zeroback/src/component/sign.css
@@ -1,8 +1,0 @@
-body {
-    margin: 0;
-    padding: 0;
-    /*background-image: url('./pulse-trace-glowing-computer-monitor-generated-by-ai.jpg');*/
-    background-size: cover;
-    background-position: center;
-    font-family: Arial, sans-serif; /* Set font family */
-}
